split to dirs
