+++ layout/document_layout.py
@@ -1,4 +1,4 @@
-from block_layout import BlockLayout
+from layout.block_layout import BlockLayout
 
 class DocumentLayout:
     HSTEP = 13
