+++ layout/draw_rect.py
